requirements.txt
Mastodon.py==1.6.3
